[OM-59747]: Fix the issue that put huge error messages in grpc calls
Intent
======
Truncate the error message bigger that 4000.

Problem
=======
An error message is shown in the log indicating that the header is too big

Diagnosis
=========
We faced a error while inserting a large number of policies. As we include name of those policies in the error message and therefore our message become bigger than 8192 size supported for the header.

Solution
========
Truncated larger message. I used 4000 as the place that we truncate the message. I believe that will have enough information for debug.