[OM-59928]: Tighten prior fix for leaked SnapshotManager resource
Tightened up prior fix for this issue. The fact the SnapshotManager
was opened in a different code block than the one in which it was
closed meant that, despite the close occurring in a `finally` clause
in a `try` block surrounding that code was insufficient. Exceptions
thrown in the interim processing would prevent the open snapshot from
being closed.

Here we move the open so that it's in the `try` block whose `finally`
clause contains the close. We should no longer see any missed closes
for these resources.