[OM-59828]: XL Repo: Reduce repository memory footprint.
Intent
=====
Lower the memory footprint of the repository's in-memory graph for large customers.

Background
==========
The in-memory graph grew in size out of proportion to the size of the topology. It was taking up a couple of hundred MB for an 80-90k topology about a year ago, and it is now taking up 1.3GB for a 430k topology. After digging around a bit, there were three significant ways to save memory:
1) Do not keep the entire uncompressed TypeSpecificInfo in memory. We only need certain parts of it, and it's very expensive to keep around - especially the volume information, which can contain hundreds of thousands of long file names/paths.
2) Do not keep duplicate strings for tag keys and values.

Also, I introduced an "ActionPartialEntity" which sends only the necessary action-related data to the action orchestrator. This was required so that we don't send the TypeSpecificInfo.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/39496/