[OM-59959]:  Fix logic when checking if group is group of clusters
Bug: When we're fetching cluster stats we have some logic to identify if the group is a cluster
or a group of clusters. In the latter case, we would expect the group being checked to have
exactly 1 expected type (one of the cluster group types), but group service adds both the cluster
type and the type of the cluster's entities, thus returning more than 1 types and making the
check fail.

Fix: Change the logic to check only the type of the direct members, instead of both
members an entities.

https://rbcommons.com/s/VMTurbo/r/40042/