[OM-60106]: Always use capacity lower bound in resize down capacity calculation. RB-40123
**The Problem**
There is a bug for containers where the market may generate actions recommending a resource limit below the request for that resource, which is not permitted by kubernetes. To prevent this, we had planned to use the capacityLowerBound. However, when the limit and request were equal, this solution doesn't work because logic in Resizer#calculateResizeDownAmount where a capacityLowerBound <= currentCapacity causes the capacityLowerBound to be ignored.

**The Solution**
The logic in Resizer#calculateResizeDownAmount was originally added to support the Resize v3 feature. Shravan later refactored the implementation to allow the market to respect capacity upper and lower bounds, but missed the update in Resizer#calculateResizeDownAmount to allow resize down to respect capacity lower bounds when current capacity is less than or equal to the current capacity lower bound. After discussion, we decided that the correct thing to do is to also fix Resizer#calculateResizeDownAmount to behave the same as Resizer#calculateResizeUpAmount.

**Testing Done**
* Updated unit tests to ensure proper respecting of capacityLowerBound for resize down actions.
* Tested this update with fix for containers to ensure limits are not sized below requests.