[OM-60191]: XL: Ignore unknown fields when deserializing JSON into Protobuf.
We dump protobufs in JSON format in multiple places in the diagnostics framework. However, when restoring diags from an older version we see errors if a field that exists in an older version got removed.

The fix is simple - we tell the parser we use to deserialize JSON into protobuf to ignore unknown fields.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/40093/