[OM-56389]: Do not overwrite the cache capacity from database when the lookback period changes
Intent
======
Prevent from having the wrong capacity for one broadcast cycle.

Problem
=======
If we have sequence of events where initially the capacity of commodity changes and after that the lookback period of that commodity changes. We will have wrong capacity for one broadcast cycle.

Diagnosis
=========
When we change the lookback period, we load data from database which has the old capacity and we will overwrite the new capacity with the old one which is wrong.

Solution
========
When the lookback period changes, use the capacity in cache as correct capacity and scale the data read from db to this capacity.