[OM-60162]: XL: Generalize DbEndpoint initialization.
DbEndpointCompleter is now a static singleton, like ComponentGrpcServer, and it gets initialized after the component context. I added a common ServerStartedNotifier utility which is used by both the ComponentGrpcServer and DbEndpointCompleter to hook into the startup of the BaseVmtComponent.

Also, the backoff iterator was failing intermittently because it was being used IN "awaitCompletion" while depending on the config property that was being set during completion. I replaced it with:
1) The default backoff strategy (every 10 seconds).
2) A maximum wait time, configurable in the "awaitCompletion" method. Tests wait up to 30 seconds. Production deployments wait up to 30 minutes.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/40064/