[OM-60310]: XL Extractor: Handle missing authRoute parameter.
Issue
=====
Extractor component never ready.

Background
==========
I added more logging, and it turns out that the extractor was failing to get the root password from the auth component. This was because the "authRoute" property is not set by default, and the logic I had added in the DbEndpointCompleter did not provide a default either.

Fix
===
Provide a default (empty) auth route - as well as other properties in the DbEndpointCompleter.
Increase logging for DB initialization.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/40162/