[OM-58952]: RB-40129 Ignore duplicate group tags when inserting in database
Intent
======
Ignore duplicate group tags when inserting in database

Problem and Diagnosis
=====================
The group upload at PwC failed because of the duplicate tags. There was tags with duplicate key-values for a single group and as the primay key for tags table is made of group id , tag key value and tag value we ended up with violating the primary key and therefore failing the insert.

The initial solution was to key ke key-value combination in a set and verify if the value is in the set before inserting in to db. However, that also failed as there was two keys which one have space at the end of it. MySQL removes the spaces from varchars when storing them (read here https://stackoverflow.com/questions/29230311/mysql-not-accepting-whitespaces). Therefore, we again end up with the same error.

Solution
========
I changed the code to use INSERT IGNORE rather than regular insert when inserting tag. This ensures we never fail insertion of tags because of duplication.