[OM-59528] RB:40197 updating unit-tests and changing decrementFunction in CommodityResizeDependencyMap to Min instead of ProjectSecond.
when the usage on the underlying provider is lower than the newCap, we should not increase the usage on the provider.
Consider a single VM on a PM. memUsed=10, old vMemCap=100, new vMemCap=13 ...
using ProjectSecond, memUsed=13
using Min, memUsed=10.

Scaling down to a capacity higher than the underlying usage on the host leads to an increase in the usage on the host without this fix.