[OM-60415]: Include the inactive scheduled policy in the policies tab
Intent
======
Include the inactive scheduled policy in the policies tab

Problem
=======
We show the policies applicable to entity in the entities tab . However, it was noticed the schedule policies are not showing in the tab when the scheduled policy is not active. That is not the case in classic and can be misleading.

Diagnosis
=========
The logic for finding associated policies to an entity goes through the different setting of the policy and gets their associated policies and use them. This is not proper approach for two reason: 1. the scheduled policies will not be shown when they are inactive. 2. if a policy loses in tie-breaking mechanism we don't show it associated to the entitiy.

Solution
========
I update the topology processor mechansim to come up with the policies associated with an entity and send them to group component as a part of entity settings. This list of policies contain all the policies that can affect the entity regardless of being active or not and we use this list to return policies associated to an entity.