[OM-60822]: Do not call getOid on null traders RB:40464
In investigating OM_60780, I noticed that some of the calls that replaced hash lookups like traderOidMap(trader) with calls to trader.getOid() happen on Trader objects that can potentially be null. The earlier version of the code is ok because the lookup into the map will simply return null if the trader is null, but the new version will throw an NPE.

After the work done in OM_60798, perform a second review of all trader.getOid() calls to ensure that none are performed on Traders that may be null.

Testing Done:
All unit tests pass.