[OM-60508]: XL AO: Keep thin topology graph in memory to speed up risk propagation.
Issue
=====
The risk propagation logic is very slow.

Background
==========
Risk propagation requires retrieving a large subset of the topology from the repository in order to see how entities connect to each other. This has proven to be quite slow, and is also an unnecessary use of resources.

The fix is to keep a local copy of the topology graph in the orchestrator, and use that copy to serve the risk propagation logic.

Implementation
==============
- Create a new 'ActionTopologyListener' which listens to the topology in the action orchestrator. The 'ActionTopologyListener' updates the 'ActionTopologyStore', creating a topology graph of 'ActionGraphEntity" objects and wrapping it in an 'ActionRealtimeTopology' object. This is the same thing we have in the repository.
- Moved the common code between AO and Repository to the topology.graph package:
    - 'BaseGraphEntity' - the actual entity. All the common properties and relationships.
    - 'BaseTopology' and 'BaseSearchableTopology'.
    - 'BaseTopologyStore' - methods to update the topology, as well as handling incremental updates.

Next Steps
==========
- Use the local graph when processing actions in the LiveActionStore (translation, description calculation, etc), and use it when possible when processing Plan actions (for entities that are NOT exclusive to the plans).

Reviewed At: https://rbcommons.com/s/VMTurbo/r/40330/