[OM-60343]: Add ON_GEN and AFTER_EXEC to ActionModeCalculator
Intent

OM-60343: Add ON_GEN and AFTER_EXEC to ActionModeCalculator so that audits are sent to ServiceNOW.

Problem

Resize up VM mem audit on fail/success is not created.

Diagnosis
Root cause is that FAILED and SUCCEEDED are not in the workflowSettingsForState map. Adding it to that map should fix the issue. ActionModeCalculator.calculateWorkflowSettings will need to be updated to include AFTER_EXECUTION workflows for FAILED and SUCCEEDED.

Solution

search for ON_GEN and AFTER_EXEC settings based on the corresponding action mode setting by using specsApplicableToAction to find the corresponding action mode settings for the current action.