[OM-60642]: Market- Translate mediation OsType license access comm to RI OSType
**Background**
It looks like we have multiple OS enums - two OsTypes in the SDK and OSType for cost probes. VMs and compute tiers use the OsType for the license access commodity, while RIs use OSType for the OS. There are a few places in the market where we properly use a map to translate between the two OS types. However, in creating the RI discounted market tier trader, we instead do a simple case-insensitive name check. This works for the base OS types, but it does not work for something like Windows with SQL standard, in which OsType defines it as Windows_SQL_Standard and OSType defines it as WINDOWS_WITH_SQL_STANDARD.

**Fix**
The fix is simple - use the same map to do the conversion as in other places within the market component. Long term, we should look at deprecating one of the enums (this is a larger change than is appropriate for a bug fix).