[OM-60846]: XL: Use component-specific DbEndpointCompleters.
Issue
=====
Extractor does not correctly initialize the database in "voltron" mode.

Background
==========
When the component starts up, the ServerStartNotifier sends the component's context to the DbEndpointCompleter, and the DbEndpointCompleter uses that context's environment to initialize the properties in the DbEndpointResolver.

The problem is, in Voltron-mode the context sent by the ServerStartNotifier is the root context (for the "voltron" super component). This means any component-specific overrides aren't present in the environment, and the "component name" in the environment is "voltron." The effect is twofold:
1) Overriding extractor-specific properties can only be done via global overrides - which is acceptable for now, but can lead to issues if we use DbEndpoints more broadly.
2) The DbEndpointResolver uses the component name as the default folder to look for db migrations, and the "voltron" sub-folder has no migrations.

Fix
===
The fix is to resurrect the SQLDatabaseConfig2 configuration class, and put the DbEndpointCompleter there. The DbEndpointCompleter is now a factory for DbEndpointBuilders. All components that want to create DbEndpoints have to import the SQLDatabaseConfig2 configuration. This way each component's context has its own copy of a DbEndpointCompleter, with the component's Environment as the source of property values.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/40477/