[OM-60971]: RI Coverage Allocator - Use applicable accounts list for RI scoping
**Background** When single scoped RI support was first added, the
coverable accounts list was not returned as part of the DTO from the
probe. Instead, consumers of the RI assumed the RI was scoped to the
purchasing account. However, Azure allows an RI to be scoped to
subscriptions other than the purchasing account so this assumption does
not hold. As part of the feature work on partial environment support,
the coverable accounts list is now part of the RI DTO. However, the RI
coverage allocator has not been updated to use this list (this is done
as part of this review).