[OM-60802]: Fix filter by "Connected to Ephemeral/Encrypted Volume" is not working (rb40645)
Description*
Problem:

"Connected to Ephemeral/Encrypted Volume" is not working.

Investigation (by Roman)

It's assumed the “ephemeral” and “encrypted” property were operating on volumes. However, we were actually hard-coding the travesal logic into the property filter definition. This is not ideal - we should use the generic approach we use for other traversal-oriented filters. Should be a small fix to the groupBuilderUseCases.json file.

Proposed fixes

Updated groupBuilderUseCases.json as suggested.

Testing Done
1. added new unit tests for the changes
2. Integration tests passed for both Ephemeral and Encrypted volume cases (see attachments).