[OM-61109]: Fix the issue that provision actions are assigned a new oid every broadcast
Intent
======
Assign same oid to provision actions

Problem
=======
PT saw a new record in ServiceNow for provision action every broadcast. Each of these rows had new action oid. This is not expected.

Diagnosis
=========
The action oid assigner creates a signature  record for each action that represents unique properties of each action. This signature and assigned oid are persisted in the DB. When new actions come in their signature is caculated and compared with existing signature. If we cannot find an existing signature we assign a new oid and save the signature. For provision actions, we used the new provision entity id as signature. This id changes every market cycle which result in new signature and consequently new oids.

Solution
========
At the current state, the provision action does not have enough information to be unique. It has the entity to clone. However, there could be multiple actions that clone a single entity. In order to address this, we add an index to each provision action which increments with new provision actions on single entity. In order word, first provision action on an entity is the provision index zero and next provision action is provision index one and so on. We use combination of source entity to clone and this index as the signature