[OM-59896]: Keep the history of capacity changes for a commodity in percentile record
Intent
======
Keep the capacity changes history for a commodity in percentile blob so we can find the index to decrement when a record goes outside of observation period

Problem
=======
The full record does not match the sum of individual record in the observation period

Diagnosis
=========
I noticed three issue when I was diagnosing this issue:
1. Currently, the index to decrement when removing a point is calculated by assuming the capacity in the oudated record, directly change to current record. This is is not accurate and may result in decrement a wrong index and eventually result in corrupted full record.

2. At the time of maintenance, we increment both latest and full record with the new observed datapoint. However, we clear the latest and then persist it and persist the full record. This will result in that last datapoint getting lost in the individual day recors.

Solution
========
The following solutions were taken:
1. We keep the history of capacity changes in the percentile records. We a point goes out of the observation period, we apply the capacity changes in order to come up with the final index to decrements.

2. We first save the daily record to previous day before saving the full record which results in clearing latests. This will result in the full record matching the daily records for an observation period.

Unit test
========
In order to make sure that the full record always matches the individual daily records. I wrote a unit test which simulates the operation of percentile editor for 60 days while the observation period is 30. At the end of simulation, it verifies that the full record is the same as aggegrating individual daily records for the loockback period.