[OM-57830]: CCA - Demand classification + RI spec matcher
**Primary Change**
The primary change in this review is to introduce demand classification for allocated demand as the subsequent stage to demand selection. Demand classification will group and classify the allocated demand as (allocated, flexibly allocated, stale, and ephemeral). Flexibly allocated demand is distinct demand from allocated demand (the latest recorded demand), but is able to covered by the same recommendation. Ephemeral demand is any demand in the duration of the mapping is under a target minimum.

**Secondary Changes**
1. Introduce a classification summary, with a detailed report (if requested). The detailed report will print the demand per (region, account, tier).
2. Remove the generic CLOUD_TIER_TYPE from all demand data structures.
3. Update the RI spec filter and matcher to use coverage scopes as a more intuitive mapping of the demand to recommendation spec.
4. Added caching to the RI spec matcher and made it available through the analysis context.
5. Add the data structures for the output of demand transformation and uncovered demand calculation.