[OM-59757]: XL: Build RedHat-derived Grafana image.
Intent
======
Build our own Grafana image, derived from redhat, to avoid security issues when deploying on-prem reporting in customer environments.

Background
==========
We are currently using the default Grafana image from Docker Hub. That image is based off of Ubuntu. We are standardized on RedHat for all the platform images that we deploy as part of our "real" installation to avoid issues with security scans. This change creates a RedHad-based Grafana image, and modifies the grafana chart to use that image.

Additional Fixes
================
Fixed a couple of issues we found with enabling XLR:
- Passing in the database password to Grafana as an environment variable. This works for special characters more reliably than the regular configmap injection.
- Linking the "grafanaAdminPassword" Spring value to the "grafana.adminPassword" specified in the CR.
- Removing the init container to dowload dashboards. We don't use that feature right now, and it's good to avoid an extra dependency.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/40805/