[OM-58759]: XL Report: Grafana user management.
Intent
======
Implement the "Phase One" option in https://vmturbo.atlassian.net/wiki/spaces/XD/pages/1575584281/Grafana+User+Management

Implementation
==============
The extractor component now creates a "report-viewer" user with "Viewer" permissions at startup (the username, password, and display name of this user is configurable). The password is a random alphanumeric string (we need some password by default).

The "UsersService" call to get the logged in user in the API component returns an X-Report-Username header with the name of the Grafana user to use for this session. The user is the username for admin users, and a shared "report-viewer" for everyone else.

Also, configuring the docker-compose version of Grafana to work with reporting.

Note - I'm not using the login token for Grafana, because the overhead of the auth check on every API call is small, and this way we don't need to deal with ttl's.

Associated API review: https://rbcommons.com/s/VMTurbo/r/40731/

Reviewed At: https://rbcommons.com/s/VMTurbo/r/40730/