[OM-61064] Added a remote writer and scrape job for telemetry.
The goal is for Prometheus to be able to collect telemetry data from XL
core components and upload them to DataCloud.

Currently for this to take effect Prometheus has to be enabled and a
DataCloud APIKEY has to be specified.

- Added a sidecar container configuration for the SevOne's Prometheus-to-
  DataCloud gateway. They don't offer some documentation document but
  the gateway accepts only a few options and they explained them to me
  so I documented them here.
- Configured prometheus to add `customer_id` and `customer_domain`
  labels to every metric that is sent to a remote writer.
- Added a remote write configuration to send just telemetry data to the
  DataCloud gateway.
- Added a scrape job for telemetry based on the 'xl' one.

Built and deployed operator then added APIKEY and verified that data is
sent to DataCloud as expected. The extra labels are there, the scrape
interval is respected, no errors were observed.

I also noticed the first upload of data is done only after the first
scrape interval has elapsed. That would mean that we'll get telemetry a
week after an instance is setup, not at the moment it is brought online,
so targets and configuration will likely be accurate by then. Not sure
if interval is reset when VM is restarted.