[OM-61604]: XL: Moving common "Pipeline" classes into a separate module.
INtent
======
Make the "Pipeline" we use in the topology processor reusable in other components.

Background
==========
Having a "Pipeline" class wrapping the operations the topology processor does to construct a topology has been very helpful as more people add more features. It provides an easy way to look at the high-level summary of what's happening, and a single access point we can use for timings, tracing, feature flags, and so on.

I'm extracting the parts of the pipeline that are not specific to the topology processor, and moving them out into a module where they can be reused by other components. Having other components use a pipeline will have two benefits:
1) If other components use the "pipeline" comments it will clean up the code and make it easier to follow and modify there as well.
2) It could make it a bit easier to move functionality out of the topology processor.

Implementation
==============
There are three abstract classes - Pipeline, PipelineContext, and PipelineSummary. Overriding them is straightforward - mainly to hide repetitive generic parameters, and provide hooks for metrics, traces, and logging.

There is also a "PipelineDefinitionBuilder" class which is used to assemble a "PipelineDefinition" required for the "Pipeline" constructor. The "PipelineDefinitionBuilder" enforces that:
1) The inputs and outputs of consecutive stages line up.
2) The input of the first stage is equal to the pipeline input.
3) The output of the last stage is equal to the pipeline output (this is actually an improvement over the current topology pipeline)

Reviewed At: https://rbcommons.com/s/VMTurbo/r/40943/