[OM-60247] Use Resize Target Utilization band for Virtual Volume commodity quantity calculation.
Requirements

In order to prevent actions that have small changes to capacities, a target utilization band is used to check if the percentile utilization falls within or outside the band. If the percentile falls within the band, then the current capacity is chosen as the resize quantity to avoid resizing by the Market. Otherwise the quantity is calculated using the resize target utilization, percentile and capacity.

Code Review: https://rbcommons.com/s/VMTurbo/r/40753/