[OM-60907]: Propagate UNKNOWN VM entity state recursivly to consumers RB:40982
**Problem**
The market is throwing a NullPointerException when attempting to propagate the effect of container resize actions through the supply chain links. We observed that this was because Containers are being sent into the market running on pods that are on VMs where the VM is in an UNKNOWN powerstate which results in the VM not being in the market. As a result, the provider of the pod is null when attempting to propagate the container resize effect to reseller providers and this causes the NPE.

**Solution**
The PropagatePowerStatePostStitchingOperation propagates the UNKNOWN entity state to consumers of VMs in POWERED_OFF or SUSPENDED powerstate. However, the operation currently does not propagate UNKNOWN to consumers of VMs in UNKNOWN state, and it also does not propagate the state recursively to consumers. The recursive part is necessary because for the typical container supply chain, we need to prevent actions all the way up the consumer chain because all those entities are affected by the uderlying provider.

**Testing Done**
* New unit tests. Existing unit tests pass.
* Reproduced the bug using diagnostics. After this change, the NPE no longer occurs.