[OM-61064] Configured DataCloud gateway to use a service account.
DataCloud service accounts are created per project from the web UI and
are represented as JSON files that contain cryptographic keys. Such a
file needs to be attached to the DataCloud gateway container's file
system to be read. Kenan Kessler from the SevOne team recommended to
use a Kubernetes secret to store the contents of this file.

I didn't find a way to create such secret from within the Prometheus
Helm chart so I'm creating it from the `xl` chart and using it in the
former.

If we find we need more telemetry-specific configuration in `xl` project
it might be worth creating a dedicated sub-chart for telemetry.

Also:

- Changed gateway's image reference to use a tag instead of the digest.
  That's mostly because it's shorter. It's also likely to have some
  meaning (like being the SHA hash of the commit that was used to build
  it) although I'm not sure which tagging policy they use in SevOne.
- Explained why `info` logging level was selected in the comments.

Built and deployed operator in an XL VM in our lab, replaced `PROJECT_ID`
and telemetry secret with those for the telemetry test project and
verified that:

- The new secret is created automatically by operator.
- The new volume is mounted into the `datacloud-gateway` container and
  no other container in the `promethes-server` pod.
- Data is sent to DataCloud successfully.