[OM-61335]: CCA - Support explicit_defaults_for_timestamp on MySQL
**Background**
With a MySQL deployment in Azure AKS, explicit_defaults_for_timestamp is enabled by default. On MySQL, unlike MariaDB, this causes any attempt to insert a NULL value for a timestamp to trigger an exception. Instead of explicitly inserting a NULL value for the entity_cloud_scope table's creation_time column, we no longer include it in the query. This allows the DB to set the value at first insertion time, while ignoring it for subsequent updates and is no longer reliant on explicit_defaults_for_timestamp being disabled for MySQL