[OM-61760]: Ensure action merge feature works on loaded diagnostics RB:41063
**Problem**
The action merge feature (ie merging actions on individual Container instances for display and execution on the workload controller that manages them) does not work when loading diagnostics.

**Solution**
It appears the reason this doesn't work is that although the ActionMergePolicyDTO information is saved in the Probes.diags file that is part of the Topology Processor diagnostics, this information is not actually restored to the ActionMergeSpecsRepository when the probe data is reloaded. The solution is simply to clear the existing data in the ActionMergeSpecsRepository and restore the saved ActionMergePolicyDTO information to it when restoring the diagnostics.

**Testing Done**
* Validated that the feature now works on loaded diags by seeing merged actions on the WorkloadController entities in the UI.
* New unit tests to ensure that the diags restore process restores to the ActionMergeSpecsRepository.