[OM-61064] Made Prometheus server enabled by default.
Since architectural constraints prevent us from turning Prometheus on
and off from within XL, Prometheus has to be always on in case telemetry
is enabled.

In order to reduce the overhead of having Prometheus enabled all the
time, I'm changing a number of satellite services to be off by default.

I'm also increasing the scrape interval of existing jobs to 1 day so
their overhead should be minimal. An alternative would be to remove
those jobs completely, but enabling them again would be more difficult
and I'm not clear on whether turbo-on-turbo and/or ARM require them.

Getting the old behaviour of Prometheus would require a section in
`charts_v1alpha1_xl_cr.yaml` similar to (note the 2 spaces before
"prometheus:"):

      prometheus:
        alertmanager:
          enabled: true
        kubeStateMetrics:
          enabled: true
        nodeExporter:
          enabled: true
        pushgateway:
          enabled: true
        server:
          global:
            scrape_interval: 1m

Built and deployed operator in an XL VM in our lab, replaced `PROJECT_ID`
and telemetry secret with those for the telemetry test project and
verified that:

- There weren't any errors in Prometheus server or DataCloud gateway
  logs (used a smaller `scrape_interval` for that).
- I could still port-forward and use the Prometheus UI.
- `storage-volume` in `prometheus-server` pod was now of `EmptyDir` type.
- The `prometheus-alertmanager`, `prometheus-kube-state-metrics`,
  `prometheus-node-exporter` and `prometheus-pushgateway` pods were gone.