[OM-60919]: XL Policy: Optimize BindToComplementaryGroup and AtMostN policy application.
Issue
=====
Auto-discovered BindToComplementaryGroup policies at BofA result in 9 million segmentation commodities.

Background
==========
A BindToComplementaryGroup policy tells the system to place a group of consumers (e.g. VMs) on any providers (e.g. Hosts) that are NOT in a particular group of providers. We implement this policy by configuring the consumers to buy a segmentation commodity, and configuring all providers in the topology EXCEPT the provider group to sell that commodity.

The CloudNative targets at BofA discover BindToComplementaryGroup policies to avoid placing CONTAINER_PODs on certain "unschedulable" nodes (VMs). We were discovering 84 of those policies at BofA. Each policy resulted in a commodity being added to almost every VM in the topology, for a total of 8.8 million sold commodities!

Fix
===
The consumer entities in a BindToComplementaryGroup policy are almost always constrained by other commodities - e.g. CLUSTER, DATACENTER - which means they cannot be placed on the vast majority of providers in a large topology. The fix is to take advantage of these existing constraints. We build up an **InvertedIndex** from all the provider entities, and use it to create segmentation commodities only on valid potential providers.

I'm applying the fix - and the same fix to the AtMostNPolicy implementation - by changing the underlying **PlacementPolicyApplication#addCommoditySoldToComplementaryProviders** method to take accept an **InvertedIndex** and a set of consumers. The **InvertedIndex** is created once for all BindToComplementaryGroup policies, and once for all AtMostN policies, and contains only entities of the relevant provider types.

Other
=====
- Modifying the **PlacementPolicyApplication** - adding a generic parameter which allows some type-friendliness in implementations.
- Modifying the **PolicyStage** summary to break down added commodities by policy type.
- Setting the display name of user-created policies properly. I wasn't able to see the policies I created in the UI without this.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41068/