[OM-61086] Added a cron job for updating Prometheus configuration.
The new job will be running every minute and it's purpose it to read
Prometheus configuration coming from a configmap and controlled by
operator, update it based on information it will get from the XL
platform and write it again in a file that will be loaded by Prometheus.

This is to avoid letting XL platform to change it's own custom resource
object.

Some work still remains to ensure container image version and configmap
name always remain consistent with the rest of XL and Prometheus
configuration. Also, it might be worth to move telemetry Kubernetes
resources to a dedicated Helm sub-chart.

Also modified prometheus-config-manager to print the contents of the
mounted configmap and verified the container indeed has access to it.