[OM-61562]: XL Report: Re-route to Turbo login page if /reports route is not authorized.
Intent
=====
Navigating to /reports route directly when not logged in should redirect to the Turbonomic login page.

Background
==========
Right now navigating to /reports when not logged in shows a generic NGINX "401" page. This looks ugly, and doesn't provide an immediate way to log in. We should re-route to the Turbo login screen, which will allow the user to log in (and then go to Reports).

The fix is trivial - we use the error_page function in NGINX to navigate to the login page. We also provide a "continue" option so that the UI re-directs the user back to the Grafana URL after login - see attached GIF.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/40973/