[OM-61086] Added a persistent volume where cron job can write the config.
Since the cron job pod needs to write the merged Prometheus configuration
and the Prometheus server pod to read it, a persistent volume was created
to be shared between the pods. Most other volume options need to share
the pods lifetime.

I couldn't find a way to mount the new persistent volume in
`prometheus-server-configmap-reload` container even though there was an
option to instruct it to monitor multiple paths. To work around that I
disabled this container altogether and added an API call to the Python
script to make the Prometheus server reload its configuration each time
the cron job is run.

Built and deployed operator in an XL VM in our lab and verified that:

- Persistent volume claim and persistent volume are created.
  In my laptop the persistent volume is created with 1MB capacity as
  expected but because of some Heketi limitation, it is created with 1GB
  capacity in my lab instance. Only a few megabytes are shown as used in
  `gluster volume status gluster-volume-path detail`.
- The `prometheus-server-configmap-reload` container isn't there any
  more.
- The configuration file is written successfully by the cron job and
  loaded by Prometheus on startup.
- Changes to the configuration are being picked up.
- Next scrape time isn't affected by reloading of configuration. i.e.
  if configuration is reloaded every minute and a job has a 5 minute
  scrape interval, it will still be scraped every 5 minutes as expected.
  Not every minute neither never.
- No errors are seen in the logs.

Notes:

- I had no success using the Helm 'template comments' documented
  [here](https://helm.sh/docs/chart_best_practices/templates/#comments-yaml-comments-vs-template-comments)
  so I sticked with the 'YAML comments' form.
- If for any reason the cron job fails to run at least once, the
  Prometheus server won't find a configuration file to read and the pod
  won't reach ready state.