[OM-61824]: Make consistent scaling group names unique RB:41140
**Problem**
Incorrect actions are being generated because consistent scaling groups with the same name are merged in the market. At the customer Cambria, there are two elasticsearch StatefulSets, one in their dev cluster and one in their build cluster. Their containers run on VMs of different sizes (the dev cluster VMs uses 8GB of VMem while the build cluster VMs have 32 GB of VMem). Due to another market bug, this results in a resize down action on the VMs in the dev cluster which would increase their utilization to over 300%. Also, because it is common for customers to run the same services in different environments, this is probably a common problem we just have not payed attention to until now.

**Solution**
When we detect scaling groups with the same name, make the name unique by injecting the OID of the scaling group into the names of the associated settings we upload to the group component. Because the scaling groups have unique OIDs, this also guarantees unique setting names.

We only do this on groups where we detect duplicates because:
1. When the setting name is already unique, there's no need to add additional information to it.
2. The longer name with the OID in it is slightly less friendly for customers to read and it is user facing because it shows up in the text of scaling actions on entities affected by the setting.
3. The longer name with the OID is more likely to hit the max length allowed for setting names in the database.

**Testing Done**
* Loaded Cambria diags. Before this fix, reproduced the bad actions observed at the customer. After this fix, the actions make much more sense.
* Looked at action descriptions. Actions on entities with duplicate consistent scaling groups now show the OID in the setting names included in the action description. Actions on entities with unique consistent scaling groups show the same name as they did before this change.
* Add unit test to test new logic.