[OM-61548]: allow removal of execution schedule from default policy
Intent
======
Allow users to remove execution schedule for default policies

Problem
=======
If the user set execution for a default policy. They will see an errror when they try to remove it. They will get an error that you cannot do that.

Diagnosis
=========
The user cannot remove default policies. We have a condition which checks if the setting has been removed from default setting and throws an exception if that happened. However, some setting can be unset by default. The examples of such settings are execution schedule.

Solution
========
There has been a recent change in default policy where we don't keep the default values in the database. Therefore, I updated the logic so if we don't provide the value for a default setting policy we check if that setting can be unset. If so, we use removed that setting value. Otherwise, we continue to use the existing value that setting policy has.